+++ src/functions.php
@@ -294,13 +294,18 @@ function wildcard(
 /**
  * Get an "empty" value for the supplied type.
  *
- * @param ReflectionType $type The type.
+ * The `$resolveSelf` parameter accepts a callback that takes no arguments, and
+ * returns the resolved class name of the `self`/`static` type.
+ *
+ * @param ReflectionType $type        The type.
+ * @param callable|null  $resolveSelf A callback to use if it is necessary to resolve a `self` or `static` type.
  *
  * @return mixed An "empty" value of the supplied type.
  */
-function emptyValue(ReflectionType $type)
+function emptyValue(ReflectionType $type, callable $resolveSelf = null)
 {
-    return Globals::$container->emptyValueFactory->fromType($type);
+    return Globals::$container->emptyValueFactory
+        ->fromType($type, $resolveSelf);
 }
 
 /**